Refactoring to remove globals for uv tools.
- global uvlayer, bm --> self.uvlayer, self.bm
- moved b_box_center to MakeIslands class
- moved rotate_island to MakeIslands class
- added get_bm method to give the rotate function a reference to the bm (instead of using globals)

Note: Global variables are not meant to be a way to move whatever around without caring about design - they are for things like the current interest rate in a banking program, or for if debug mode is on in a program. The general rule of thumb is that you NEVER change the value of global variables and the name is always in all caps to show this. The only time you MIGHT change its value is if the user can explicitly change the setting in the UI.