Added progress printouts, ascii art logo.
I used an ascii generator, then made the 'T' into a tree by hand. This is displayed every time the tree is generated and when the add-on is loaded. BTW the 'r' before the string for the logo means to interpret it as a raw string (the '\' will be displayed as such and not used as a "special" character).