Added force field effect
I couldn't find a way to acces the scene field data (the direction and
strength of the net force field at any point of the scene) so I was just
able to take the wind and point force into account.

The position and rotation of the tree isn't applied yet so it only works
if the tree is not rotated and is at the origin point.